refactor: delete legacy composite.py / composite_*.py / generate*.py files
The standardized v2ecoli.composites + v2ecoli.core framework fully replaces
the six legacy files. Every caller has been migrated. Shared helpers (make_edge,
inject_flow_dependencies, _seed_state_from_defaults, _seed_mass_listener,
_normalize_boundary_units, _make_instance, _get_special_step, _expand_flushes,
FLUSH, PARTITIONED_PROCESSES, ALL_PARTITIONED, ALLOCATOR_LAYERS) relocated to
v2ecoli/composites/_helpers.py. Architecture-specific helpers
(build_execution_layers, DEFAULT_FEATURES, _get_step_config) inlined into
the corresponding composites/<arch>.py files.

Also removes stale 'composite.py' entry from the pickle allowlist in
tests/test_pickle_allowlist.py — the file no longer exists.

This commit completes the atomic refactor — there is no half-state where
both APIs are usable.
