+++ tests/test_pickle_allowlist.py
@@ -33,8 +33,6 @@ BANNED = frozenset({'pickle', 'dill', 'cloudpickle'})
 
 # File-level allowlist (paths relative to v2ecoli/). Value explains why.
 ALLOWLIST: dict[str, str] = {
-    'composite.py':
-        'loads ParCa sim_data_cache.dill (ParCa cache — AGENTS.md:114 exception)',
     'core.py':
         'shared cache helpers extracted from composite.py — same ParCa dill exception',
     'library/sim_data.py':
