+++ v2ecoli/composites/baseline.py
@@ -5,21 +5,20 @@ vivarium-collective/vEcoli composite tick-for-tick. See AGENTS.md.
 
 Migration note: the document-building body was migrated from
 ``v2ecoli/generate.py:build_document`` and
-``v2ecoli/composite.py:_build_from_cache``.  Both legacy files still exist
-and are unchanged; they will be deleted in Task 14.
+``v2ecoli/composite.py:_build_from_cache``.  Both legacy files were deleted
+in Task 14.
 
-Helper functions (``make_edge``, ``inject_flow_dependencies``,
+Shared helpers (``make_edge``, ``inject_flow_dependencies``,
 ``_seed_state_from_defaults``, ``_seed_mass_listener``,
 ``_normalize_boundary_units``, ``_make_instance``, ``_get_special_step``,
-``_expand_flushes``, module-level constants) are imported from
-``v2ecoli.generate`` rather than duplicated here because
-``generate_departitioned.py`` and ``generate_reconciled.py`` import the same
-helpers.  Task 4 will resolve the shared-helpers concern when it migrates
-those two files.
+module-level constants) live in ``v2ecoli.composites._helpers``.
+Architecture-specific helpers (``build_execution_layers``, ``DEFAULT_FEATURES``,
+``_get_step_config``) are defined as private module-level functions here.
 """
 
 from __future__ import annotations
 
+import copy
 from typing import Any
 
 import numpy as np
@@ -29,32 +28,314 @@ from pbg_superpowers.composite_generator import composite_generator
 from v2ecoli.core import build_core, load_cache_bundle
 
 # ---------------------------------------------------------------------------
-# Re-use shared helpers and constants from generate.py.
-# generate_departitioned.py and generate_reconciled.py import the same names,
-# so duplication here would conflict in Task 14 when generate.py is deleted.
-# Task 4 is responsible for deciding the final home for these helpers.
+# Shared helpers and constants
 # ---------------------------------------------------------------------------
-from v2ecoli.generate import (
-    # helpers
+from v2ecoli.composites._helpers import (
     make_edge,
     inject_flow_dependencies,
     _seed_state_from_defaults,
     _seed_mass_listener,
     _normalize_boundary_units,
-    _get_step_config,
+    _make_instance,
+    _get_special_step,
     _expand_flushes,
-    build_execution_layers,
-    # constants
     FLUSH,
     PARTITIONED_PROCESSES,
     ALL_PARTITIONED,
-    DEFAULT_FEATURES,
+    ALLOCATOR_LAYERS,
 )
 
+
+# ---------------------------------------------------------------------------
+# Execution layers (partitioned / baseline architecture)
+# ---------------------------------------------------------------------------
+
+BASE_EXECUTION_LAYERS = [
+    # Layer 0: post-division mass
+    ['post-division-mass-listener'], FLUSH,
+
+    # Layer 1: media/environment (sequential sub-steps)
+    ['media_update'], FLUSH,
+    ['ecoli-tf-unbinding'],
+    ['exchange_data'], FLUSH,
+
+    # Layer 2: standalone (no partitioning needed)
+    ['ecoli-equilibrium', 'ecoli-two-component-system', 'ecoli-rna-maturation'], FLUSH,
+
+    # Layer 3: TF binding
+    ['ecoli-tf-binding'], FLUSH,
+
+    # Layer 4: protein degradation (standalone — no resource competition)
+    ['ecoli-protein-degradation'],
+
+    # Layer 4b: standalone initiation/replication/complexation
+    ['ecoli-complexation', 'ecoli-chromosome-replication',
+     'ecoli-polypeptide-initiation', 'ecoli-transcript-initiation'],
+    # RNA degradation still partitioned (shares water with other processes)
+    ['ecoli-rna-degradation_requester'],
+    ['allocator_2'],
+    ['ecoli-rna-degradation_evolver'], FLUSH,
+
+    # Layer 5: partition layer 3 -- elongation requesters (parallel)
+    ['ecoli-polypeptide-elongation_requester', 'ecoli-transcript-elongation_requester'],
+    ['allocator_3'],
+    # Layer 5: partition layer 3 -- elongation evolvers (parallel)
+    ['ecoli-polypeptide-elongation_evolver', 'ecoli-transcript-elongation_evolver'], FLUSH,
+
+    # Layer 6: chromosome structure + metabolism (sequential)
+    ['ecoli-chromosome-structure'], FLUSH,
+    ['ecoli-metabolism'], FLUSH,
+
+    # Layer 7: listeners (parallel)
+    ['RNA_counts_listener', 'ecoli-mass-listener',
+     'monomer_counts_listener', 'replication_data_listener', 'ribosome_data_listener',
+     'rna_synth_prob_listener', 'rnap_data_listener', 'unique_molecule_counts'], FLUSH,
+
+    # Emitter + clock
+    ['emitter'],
+    ['global_clock'],
+
+    # Layer 8: division check
+    ['mark_d_period'], FLUSH,
+    ['division'],
+]
+
+FEATURE_MODULES = {
+    'supercoiling': {
+        'insert_after': 'ecoli-chromosome-structure',
+        'steps': ['dna-supercoiling-step'],
+        'listeners': ['dna_supercoiling_listener'],
+    },
+    'ppgpp_regulation': {
+        'insert_before': 'ecoli-transcript-initiation',
+        'steps': ['ppgpp-initiation'],
+    },
+    'trna_attenuation': {
+        'insert_before': 'ecoli-transcript-elongation_requester',
+        'steps': ['trna-attenuation-config'],
+    },
+}
+
+DEFAULT_FEATURES = ['ppgpp_regulation']  # trna_attenuation disabled to match v1 default
+
+
+def build_execution_layers(features=None):
+    """Build EXECUTION_LAYERS from base + enabled feature modules."""
+    layers = copy.deepcopy(BASE_EXECUTION_LAYERS)
+    for feat_name in (features or []):
+        feat = FEATURE_MODULES.get(feat_name)
+        if feat is None:
+            continue
+        if 'insert_after' in feat:
+            ref = feat['insert_after']
+            for i, layer in enumerate(layers):
+                if isinstance(layer, list) and ref in layer:
+                    for step_name in feat.get('steps', []):
+                        layers.insert(i + 1, [step_name])
+                    break
+        if 'insert_before' in feat:
+            ref = feat['insert_before']
+            for i, layer in enumerate(layers):
+                if isinstance(layer, list) and ref in layer:
+                    for step_name in reversed(feat.get('steps', [])):
+                        layers.insert(i, [step_name])
+                    break
+        for listener in feat.get('listeners', []):
+            for layer in layers:
+                if isinstance(layer, list) and 'ecoli-mass-listener' in layer:
+                    if listener not in layer:
+                        layer.append(listener)
+                    break
+    return _expand_flushes(layers)
+
+
 # Convenience re-export: ordered list of all step names in execution order.
-# Computed once at module import so callers don't need to import from
-# v2ecoli.generate directly.
 FLOW_ORDER = [step for layer in build_execution_layers(DEFAULT_FEATURES) for step in layer]
+
+
+# ---------------------------------------------------------------------------
+# Step instantiation (partitioned / baseline architecture)
+# ---------------------------------------------------------------------------
+
+def _get_step_config(loader, step_name, core, process_cache=None):
+    """Get (instance, topology, edge_type[, in_topo, out_topo]) for a step."""
+    from v2ecoli.processes.equilibrium import Equilibrium
+    from v2ecoli.processes.two_component_system import TwoComponentSystem
+    from v2ecoli.processes.rna_maturation import RnaMaturation
+    from v2ecoli.processes.complexation import Complexation
+    from v2ecoli.processes.protein_degradation import ProteinDegradation
+    from v2ecoli.processes.rna_degradation import RnaDegradation
+    from v2ecoli.processes.transcript_initiation import TranscriptInitiation
+    from v2ecoli.processes.transcript_elongation import TranscriptElongation
+    from v2ecoli.processes.polypeptide_initiation import PolypeptideInitiation
+    from v2ecoli.processes.polypeptide_elongation import PolypeptideElongation
+    from v2ecoli.processes.chromosome_replication import ChromosomeReplication
+    from v2ecoli.processes.tf_binding import TfBinding
+    from v2ecoli.processes.tf_unbinding import TfUnbinding
+    from v2ecoli.processes.chromosome_structure import ChromosomeStructure
+    from v2ecoli.processes.metabolism import Metabolism
+    from v2ecoli.steps.partition import Requester, Evolver
+    from v2ecoli.steps.listeners.mass_listener import MassListener, PostDivisionMassListener
+    from v2ecoli.steps.listeners.rna_counts import RNACounts
+    from v2ecoli.steps.listeners.rna_synth_prob import RnaSynthProb
+    from v2ecoli.steps.listeners.monomer_counts import MonomerCounts
+    from v2ecoli.steps.listeners.dna_supercoiling import DnaSupercoiling
+    from v2ecoli.steps.listeners.replication_data import ReplicationData
+    from v2ecoli.steps.listeners.rnap_data import RnapData
+    from v2ecoli.steps.listeners.unique_molecule_counts import UniqueMoleculeCounts
+    from v2ecoli.steps.listeners.ribosome_data import RibosomeData
+    from v2ecoli.steps.media_update import MediaUpdate
+    from v2ecoli.steps.exchange_data import ExchangeData
+
+    if process_cache is None:
+        process_cache = {}
+
+    base_name = step_name.replace('_requester', '').replace('_evolver', '')
+
+    # Handle allocators
+    if step_name.startswith('allocator'):
+        from v2ecoli.steps.allocator import Allocator
+        try:
+            alloc_config = loader.get_config_by_name('allocator')
+        except Exception:
+            alloc_config = {}
+        layer_procs = ALLOCATOR_LAYERS.get(step_name, ALL_PARTITIONED)
+        if not alloc_config.get('process_names'):
+            alloc_config['process_names'] = ALL_PARTITIONED
+        alloc_config['layer_processes'] = layer_procs
+        instance = _make_instance(Allocator, alloc_config, core)
+        topo = instance.topology
+        return instance, topo, 'step'
+
+    try:
+        config = loader.get_config_by_name(base_name)
+    except (KeyError, AttributeError):
+        try:
+            config = loader.get_config_by_name(step_name)
+        except (KeyError, AttributeError):
+            return _get_special_step(loader, step_name, core)
+
+    if config is None:
+        return _get_special_step(loader, step_name, core)
+
+    # _instantiate_step inlined here (baseline/partitioned version)
+    STANDALONE_STEPS = {
+        'ecoli-tf-binding': TfBinding,
+        'ecoli-tf-unbinding': TfUnbinding,
+        'ecoli-chromosome-structure': ChromosomeStructure,
+        'ecoli-metabolism': Metabolism,
+        'ecoli-protein-degradation': ProteinDegradation,
+        'ecoli-equilibrium': Equilibrium,
+        'ecoli-two-component-system': TwoComponentSystem,
+        'ecoli-complexation': Complexation,
+        'ecoli-rna-maturation': RnaMaturation,
+        'ecoli-transcript-initiation': TranscriptInitiation,
+        'ecoli-polypeptide-initiation': PolypeptideInitiation,
+        'ecoli-chromosome-replication': ChromosomeReplication,
+    }
+
+    SIMPLE_STEPS = {
+        'ecoli-mass-listener': MassListener,
+        'post-division-mass-listener': PostDivisionMassListener,
+        'RNA_counts_listener': RNACounts,
+        'rna_synth_prob_listener': RnaSynthProb,
+        'monomer_counts_listener': MonomerCounts,
+        'dna_supercoiling_listener': DnaSupercoiling,
+        'replication_data_listener': ReplicationData,
+        'rnap_data_listener': RnapData,
+        'unique_molecule_counts': UniqueMoleculeCounts,
+        'ribosome_data_listener': RibosomeData,
+        'media_update': MediaUpdate,
+        'exchange_data': ExchangeData,
+    }
+
+    from v2ecoli.library.config_resolver import resolve_config
+    config = resolve_config(config) if config else config
+
+    # Partitioned processes: wrap with generic Requester/Evolver
+    if base_name in PARTITIONED_PROCESSES:
+        proc_cls = PARTITIONED_PROCESSES[base_name]
+        if base_name in process_cache:
+            process = process_cache[base_name]
+        else:
+            from v2ecoli.library.ecoli_step import set_current_core
+            set_current_core(core)
+            process = proc_cls(config)
+            set_current_core(None)
+            process_cache[base_name] = process
+        topology = dict(config.get('topology', {}) or {})
+        if not topology:
+            topology = getattr(process, 'topology',
+                               getattr(proc_cls, 'topology', {}))
+            if callable(topology):
+                topology = topology()
+            topology = dict(topology)
+
+        if step_name.endswith('_requester'):
+            instance = Requester({
+                'time_step': config.get('time_step', 1),
+                'process': process,
+            })
+            in_topo = dict(topology)
+            in_topo['global_time'] = ('global_time',)
+            in_topo.setdefault('timestep', ('timestep',))
+            in_topo['next_update_time'] = ('next_update_time', base_name)
+            in_topo['process'] = ('process', base_name)
+            out_ports = set(instance.outputs().keys())
+            out_topo = {
+                'next_update_time': ('next_update_time', base_name),
+                'process': ('process', base_name),
+            }
+            if 'request' in out_ports:
+                out_topo['request'] = ('request', base_name)
+            if 'listeners' in out_ports:
+                out_topo['listeners'] = topology.get('listeners', ('listeners',))
+            return instance, topology, 'step', in_topo, out_topo
+
+        elif step_name.endswith('_evolver'):
+            instance = Evolver({
+                'time_step': config.get('time_step', 1),
+                'process': process,
+            })
+            in_topo = dict(topology)
+            in_topo['allocate'] = ('allocate', base_name)
+            in_topo['global_time'] = ('global_time',)
+            in_topo.setdefault('timestep', ('timestep',))
+            in_topo['next_update_time'] = ('next_update_time', base_name)
+            in_topo['process'] = ('process', base_name)
+            out_ports = set(instance.outputs().keys())
+            out_topo = {
+                'next_update_time': ('next_update_time', base_name),
+                'process': ('process', base_name),
+            }
+            for port in out_ports:
+                if port in ('next_update_time', 'process', 'allocate',
+                            'global_time', 'timestep'):
+                    continue
+                if port in topology:
+                    out_topo[port] = topology[port]
+                elif port == 'listeners':
+                    out_topo['listeners'] = ('listeners',)
+            return instance, topology, 'step', in_topo, out_topo
+
+    # Standalone steps
+    if step_name in STANDALONE_STEPS:
+        step_cls = STANDALONE_STEPS[step_name]
+        instance = _make_instance(step_cls, config, core)
+        topology = getattr(instance, 'topology', {})
+        if callable(topology):
+            topology = topology()
+        return instance, topology, 'step'
+
+    elif step_name in SIMPLE_STEPS:
+        cls = SIMPLE_STEPS[step_name]
+        instance = _make_instance(cls, config, core)
+        topology = getattr(instance, 'topology', {})
+        if callable(topology):
+            topology = topology()
+        return instance, topology, 'step'
+
+    return None
 
 
 @composite_generator(
@@ -82,7 +363,7 @@ def baseline(core: Any = None, *, seed: int = 0, cache_dir: str = "out/cache") -
 
     Note: ``features`` is fixed to ``DEFAULT_FEATURES`` and is not a caller-
     visible parameter.  To run with a different feature set, use the
-    departitioned or reconciled generator (when they land).
+    departitioned or reconciled generator.
 
     Args:
         core: bigraph-schema core.  If None, one is created via build_core().
@@ -98,18 +379,11 @@ def baseline(core: Any = None, *, seed: int = 0, cache_dir: str = "out/cache") -
         core = build_core()
 
     bundle = load_cache_bundle(cache_dir)
-    # bundle is a flat dict: initial_state, configs, unique_names,
-    # dry_mass_inc_dict (and possibly others from the dill cache).
     initial_state = bundle["initial_state"]
     configs = bundle["configs"]
     unique_names = bundle["unique_names"]
     dry_mass_inc_dict = bundle.get("dry_mass_inc_dict", {})
 
-    # ------------------------------------------------------------------
-    # Body migrated verbatim from v2ecoli/generate.py:build_document
-    # (signature arguments replaced with bundle keys above).
-    # features defaults to DEFAULT_FEATURES (same as build_document).
-    # ------------------------------------------------------------------
     features = DEFAULT_FEATURES
 
     cell_state = {}
@@ -141,25 +415,21 @@ def baseline(core: Any = None, *, seed: int = 0, cache_dir: str = "out/cache") -
     })
 
     # Initialize next_update_time for all partitioned processes
-    # (requesters/evolvers check this to decide whether to run)
     nut = cell_state.setdefault('next_update_time', {})
     for proc_name in ALL_PARTITIONED:
-        nut.setdefault(proc_name, 0.0)  # Run on first tick
+        nut.setdefault(proc_name, 0.0)
 
-    # Pre-create shared request/allocate/process stores with per-process sub-keys
+    # Pre-create shared request/allocate/process stores
     cell_state.setdefault('request', {})
     cell_state.setdefault('allocate', {})
     for proc_name in ALL_PARTITIONED:
         cell_state['request'].setdefault(proc_name, {'bulk': {}})
         cell_state['allocate'].setdefault(proc_name, {'bulk': {}})
-        # process store starts empty — populated when requester runs
-    # Pre-create listener sub-stores expected by various steps
     n_part = len(ALL_PARTITIONED)
     cell_state['listeners'].setdefault('atp', {
         'atp_requested': np.zeros(n_part, dtype=int),
         'atp_allocated_initial': np.zeros(n_part, dtype=int),
     })
-    # Listener sub-stores are populated by _seed_state_from_ports below
 
     cell_state.setdefault('process_state', {})
     cell_state['process_state'].setdefault('polypeptide_elongation', {
@@ -218,7 +488,6 @@ def baseline(core: Any = None, *, seed: int = 0, cache_dir: str = "out/cache") -
                     instance, topology, edge_type=edge_type)
 
     # Place shared PartitionedProcess instances in the process store
-    # (Requester/Evolver read them from state via the 'process' port)
     for proc_name, proc_instance in _process_cache.items():
         cell_state['process'][proc_name] = (proc_instance,)
 
@@ -239,5 +508,3 @@ def baseline(core: Any = None, *, seed: int = 0, cache_dir: str = "out/cache") -
         'sequential_steps': False,
         'flow_order': flow_order,
     }
-
-
