+++ v2ecoli/composites/departitioned.py
@@ -3,18 +3,18 @@
 41 steps with requester+evolver halves fused into single Steps. See AGENTS.md.
 
 Body migrated from:
-  - v2ecoli/generate_departitioned.py:build_departitioned_document — top-level document builder
-  - v2ecoli/composite_departitioned.py:make_departitioned_composite — thin Composite wrapper
-    (no meaningful post-processing; just passes args through to build_departitioned_document)
+  - v2ecoli/generate_departitioned.py:build_departitioned_document
+  - v2ecoli/composite_departitioned.py:make_departitioned_composite
 
-Helpers from v2ecoli.generate (and architecture-specific helpers from
-v2ecoli.generate_departitioned) are imported rather than duplicated here.
-Task 14's deletion of the legacy generate*.py files will need to resolve
-where the shared helpers ultimately live.
+Shared helpers live in ``v2ecoli.composites._helpers``.
+Architecture-specific helpers (``build_execution_layers``, ``DEFAULT_FEATURES``,
+``_get_step_config``) are defined as private module-level functions here.
+Both legacy files were deleted in Task 14.
 """
 
 from __future__ import annotations
 
+import copy
 from typing import Any
 
 import numpy as np
@@ -24,24 +24,277 @@ from pbg_superpowers.composite_generator import composite_generator
 from v2ecoli.core import build_core, load_cache_bundle
 
 # ---------------------------------------------------------------------------
-# Re-use shared helpers and constants from generate.py
-# (same set that generate_departitioned.py imports)
+# Shared helpers and constants
 # ---------------------------------------------------------------------------
-from v2ecoli.generate import (
+from v2ecoli.composites._helpers import (
     make_edge,
     inject_flow_dependencies,
     _seed_state_from_defaults,
     _seed_mass_listener,
     _normalize_boundary_units,
+    _make_instance,
+    _get_special_step,
+    _expand_flushes,
+    FLUSH,
+    PARTITIONED_PROCESSES,
     ALL_PARTITIONED,
 )
 
-# Architecture-specific helpers from generate_departitioned.py
-from v2ecoli.generate_departitioned import (
-    build_execution_layers,
-    _get_step_config,
-    DEFAULT_FEATURES,
-)
+# Process imports needed by _instantiate_departitioned_step
+from v2ecoli.processes.equilibrium import Equilibrium
+from v2ecoli.processes.two_component_system import TwoComponentSystem
+from v2ecoli.processes.rna_maturation import RnaMaturation
+from v2ecoli.processes.complexation import Complexation
+from v2ecoli.processes.protein_degradation import ProteinDegradation
+from v2ecoli.processes.transcript_initiation import TranscriptInitiation
+from v2ecoli.processes.polypeptide_initiation import PolypeptideInitiation
+from v2ecoli.processes.chromosome_replication import ChromosomeReplication
+from v2ecoli.steps.departitioned import DepartitionedStep
+
+
+# ---------------------------------------------------------------------------
+# Feature modules (departitioned step names)
+# ---------------------------------------------------------------------------
+
+FEATURE_MODULES = {
+    'supercoiling': {
+        'insert_after': 'ecoli-chromosome-structure',
+        'steps': ['dna-supercoiling-step'],
+        'listeners': ['dna_supercoiling_listener'],
+    },
+    'ppgpp_regulation': {
+        'insert_before': 'ecoli-transcript-initiation',
+        'steps': ['ppgpp-initiation'],
+    },
+    'trna_attenuation': {
+        'insert_before': 'ecoli-transcript-elongation',
+        'steps': ['trna-attenuation-config'],
+    },
+}
+
+DEFAULT_FEATURES = ['ppgpp_regulation']  # trna_attenuation disabled to match v1 default
+
+
+# ---------------------------------------------------------------------------
+# Execution layers (departitioned — no allocators)
+# ---------------------------------------------------------------------------
+
+BASE_EXECUTION_LAYERS = [
+    # Layer 0: post-division mass
+    ['post-division-mass-listener'], FLUSH,
+
+    # Layer 1: media/environment
+    ['media_update'], FLUSH,
+    ['ecoli-tf-unbinding'],
+    ['exchange_data'], FLUSH,
+
+    # Layer 2: partition layer 1 (sequential standalone)
+    ['ecoli-equilibrium'],
+    ['ecoli-rna-maturation'],
+    ['ecoli-two-component-system'], FLUSH,
+
+    # Layer 3: TF binding
+    ['ecoli-tf-binding'], FLUSH,
+
+    # Layer 4: partition layer 2 (sequential standalone)
+    ['ecoli-complexation'],
+    ['ecoli-protein-degradation'],
+    ['ecoli-rna-degradation'],
+    ['ecoli-transcript-initiation'],
+    ['ecoli-polypeptide-initiation'],
+    ['ecoli-chromosome-replication'], FLUSH,
+
+    # Layer 5: partition layer 3 (sequential standalone)
+    ['ecoli-transcript-elongation'],
+    ['ecoli-polypeptide-elongation'], FLUSH,
+
+    # Layer 6: chromosome structure + metabolism
+    ['ecoli-chromosome-structure'], FLUSH,
+    ['ecoli-metabolism'], FLUSH,
+
+    # Layer 7: listeners (parallel)
+    ['RNA_counts_listener', 'dna_supercoiling_listener', 'ecoli-mass-listener',
+     'monomer_counts_listener', 'replication_data_listener', 'ribosome_data_listener',
+     'rna_synth_prob_listener', 'rnap_data_listener', 'unique_molecule_counts'], FLUSH,
+
+    # Emitter + clock
+    ['emitter'],
+    ['global_clock'],
+
+    # Layer 8: division check
+    ['mark_d_period'], FLUSH,
+    ['division'],
+]
+
+
+def build_execution_layers(features=None):
+    """Build EXECUTION_LAYERS from base + enabled feature modules."""
+    layers = copy.deepcopy(BASE_EXECUTION_LAYERS)
+    for feat_name in (features or []):
+        feat = FEATURE_MODULES.get(feat_name)
+        if feat is None:
+            continue
+        if 'insert_after' in feat:
+            ref = feat['insert_after']
+            for i, layer in enumerate(layers):
+                if isinstance(layer, list) and ref in layer:
+                    for step_name in feat.get('steps', []):
+                        layers.insert(i + 1, [step_name])
+                    break
+        if 'insert_before' in feat:
+            ref = feat['insert_before']
+            for i, layer in enumerate(layers):
+                if isinstance(layer, list) and ref in layer:
+                    for step_name in reversed(feat.get('steps', [])):
+                        layers.insert(i, [step_name])
+                    break
+        for listener in feat.get('listeners', []):
+            for layer in layers:
+                if isinstance(layer, list) and 'ecoli-mass-listener' in layer:
+                    if listener not in layer:
+                        layer.append(listener)
+                    break
+    return _expand_flushes(layers)
+
+
+# ---------------------------------------------------------------------------
+# Step instantiation (departitioned)
+# ---------------------------------------------------------------------------
+
+def _instantiate_departitioned_step(step_name, config, loader, core):
+    """Instantiate a step for the departitioned architecture."""
+    from v2ecoli.processes.tf_binding import TfBinding
+    from v2ecoli.processes.tf_unbinding import TfUnbinding
+    from v2ecoli.processes.chromosome_structure import ChromosomeStructure
+    from v2ecoli.processes.metabolism import Metabolism
+    from v2ecoli.steps.listeners.mass_listener import (
+        MassListener, PostDivisionMassListener)
+    from v2ecoli.steps.listeners.rna_counts import RNACounts
+    from v2ecoli.steps.listeners.rna_synth_prob import RnaSynthProb
+    from v2ecoli.steps.listeners.monomer_counts import MonomerCounts
+    from v2ecoli.steps.listeners.dna_supercoiling import DnaSupercoiling
+    from v2ecoli.steps.listeners.replication_data import ReplicationData
+    from v2ecoli.steps.listeners.rnap_data import RnapData
+    from v2ecoli.steps.listeners.unique_molecule_counts import UniqueMoleculeCounts
+    from v2ecoli.steps.listeners.ribosome_data import RibosomeData
+    from v2ecoli.steps.media_update import MediaUpdate
+    from v2ecoli.steps.exchange_data import ExchangeData
+
+    STANDALONE_STEPS = {
+        'ecoli-tf-binding': TfBinding,
+        'ecoli-tf-unbinding': TfUnbinding,
+        'ecoli-chromosome-structure': ChromosomeStructure,
+        'ecoli-metabolism': Metabolism,
+        'ecoli-equilibrium': Equilibrium,
+        'ecoli-two-component-system': TwoComponentSystem,
+        'ecoli-rna-maturation': RnaMaturation,
+        'ecoli-complexation': Complexation,
+        'ecoli-protein-degradation': ProteinDegradation,
+        'ecoli-transcript-initiation': TranscriptInitiation,
+        'ecoli-polypeptide-initiation': PolypeptideInitiation,
+        'ecoli-chromosome-replication': ChromosomeReplication,
+    }
+
+    SIMPLE_STEPS = {
+        'ecoli-mass-listener': MassListener,
+        'post-division-mass-listener': PostDivisionMassListener,
+        'RNA_counts_listener': RNACounts,
+        'rna_synth_prob_listener': RnaSynthProb,
+        'monomer_counts_listener': MonomerCounts,
+        'dna_supercoiling_listener': DnaSupercoiling,
+        'replication_data_listener': ReplicationData,
+        'rnap_data_listener': RnapData,
+        'unique_molecule_counts': UniqueMoleculeCounts,
+        'ribosome_data_listener': RibosomeData,
+        'media_update': MediaUpdate,
+        'exchange_data': ExchangeData,
+    }
+
+    from v2ecoli.library.config_resolver import resolve_config
+    config = resolve_config(config) if config else config
+
+    EVOLVE_ONLY = {
+        'ecoli-rna-maturation',
+        'ecoli-complexation',
+    }
+
+    # Departitioned: wrap PartitionedProcess in DepartitionedStep
+    if step_name in PARTITIONED_PROCESSES:
+        proc_cls = PARTITIONED_PROCESSES[step_name]
+        from v2ecoli.library.ecoli_step import set_current_core
+        set_current_core(core)
+        process = proc_cls(config)
+        set_current_core(None)
+
+        if step_name in EVOLVE_ONLY:
+            process.evolve_only = True
+
+        topology = dict(config.get('topology', {}) or {})
+        if not topology:
+            topology = getattr(process, 'topology',
+                               getattr(proc_cls, 'topology', {}))
+            if callable(topology):
+                topology = topology()
+            topology = dict(topology)
+
+        instance = DepartitionedStep({
+            'time_step': config.get('time_step', 1),
+            'process': process,
+        })
+
+        in_topo = dict(topology)
+        in_topo['global_time'] = ('global_time',)
+        in_topo.setdefault('timestep', ('timestep',))
+        in_topo['next_update_time'] = ('next_update_time', step_name)
+
+        out_topo = {
+            'next_update_time': ('next_update_time', step_name),
+        }
+        out_ports = set(instance.outputs().keys())
+        for port in out_ports:
+            if port in ('next_update_time', 'global_time', 'timestep'):
+                continue
+            if port in topology:
+                out_topo[port] = topology[port]
+            elif port == 'listeners':
+                out_topo['listeners'] = topology.get(
+                    'listeners', ('listeners',))
+
+        return instance, topology, 'step', in_topo, out_topo
+
+    if step_name in STANDALONE_STEPS:
+        step_cls = STANDALONE_STEPS[step_name]
+        instance = _make_instance(step_cls, config, core)
+        topology = getattr(instance, 'topology', {})
+        if callable(topology):
+            topology = topology()
+        return instance, topology, 'step'
+
+    if step_name in SIMPLE_STEPS:
+        cls = SIMPLE_STEPS[step_name]
+        instance = _make_instance(cls, config, core)
+        topology = getattr(instance, 'topology', {})
+        if callable(topology):
+            topology = topology()
+        return instance, topology, 'step'
+
+    return None
+
+
+def _get_step_config(loader, step_name, core):
+    """Get step config — skip allocators."""
+    if step_name.startswith('allocator'):
+        return None
+
+    try:
+        config = loader.get_config_by_name(step_name)
+    except (KeyError, AttributeError):
+        return _get_special_step(loader, step_name, core)
+
+    if config is None:
+        return _get_special_step(loader, step_name, core)
+
+    return _instantiate_departitioned_step(step_name, config, loader, core)
 
 
 @composite_generator(
@@ -72,8 +325,7 @@ def departitioned(core: Any = None, *, seed: int = 0, cache_dir: str = "out/cach
 
     Note: ``seed`` is accepted for signature uniformity across all three
     architectures, but the departitioned document is deterministic; the
-    seed is not consumed during construction. (Stochastic processes inside
-    the composite use the seed during simulation, not at build time.)
+    seed is not consumed during construction.
 
     Args:
         core: bigraph-schema core. If None, one is created via build_core().
@@ -90,19 +342,10 @@ def departitioned(core: Any = None, *, seed: int = 0, cache_dir: str = "out/cach
         core = build_core()
 
     bundle = load_cache_bundle(cache_dir)
-    # bundle is a flat dict: initial_state, configs, unique_names,
-    # dry_mass_inc_dict (and possibly others from the dill cache).
     initial_state = bundle["initial_state"]
     configs = bundle["configs"]
     unique_names = bundle["unique_names"]
     dry_mass_inc_dict = bundle.get("dry_mass_inc_dict", {})
-
-    # ------------------------------------------------------------------
-    # Body migrated verbatim from
-    # v2ecoli/generate_departitioned.py:build_departitioned_document
-    # (signature arguments replaced with bundle keys above).
-    # features defaults to DEFAULT_FEATURES.
-    # ------------------------------------------------------------------
 
     cell_state = {}
     cell_state.update(initial_state)
@@ -188,8 +431,6 @@ def departitioned(core: Any = None, *, seed: int = 0, cache_dir: str = "out/cach
     for step_name in flow_order:
         config = _get_step_config(loader, step_name, core)
         if config is None:
-            # Match baseline behavior: silently skip steps without a resolvable
-            # config (e.g. exchange_data, which has no LoadSimData entry).
             continue
         if len(config) == 5:
             instance, topology, edge_type, in_topo, out_topo = config
